move generic components for all layouts
